Fix implicit float-to-int conversion warnings
Add explicit casts to silence -Wfloat-conversion warnings:
- htmath.c: cast math rendering coordinates to int
- html.c: cast HTCoord style properties to int, use explicit != 0 for bool check
- rle.c: cast RGB-to-grayscale result to unsigned char
- window.c: cast window dimension limits to unsigned int

+++ src/libIMG/window.c
@@ -548,12 +548,12 @@ XImageInfo* imageInWindow(Display* disp,
         if (!winwidth) {
             winwidth = image->width;
             if (winwidth > DisplayWidth(disp, scrn) * 0.9)
-                winwidth = DisplayWidth(disp, scrn) * 0.9;
+                winwidth = (unsigned int)(DisplayWidth(disp, scrn) * 0.9);
         }
         if (!winheight) {
             winheight = image->height;
             if (winheight > DisplayHeight(disp, scrn) * 0.9)
-                winheight = DisplayHeight(disp, scrn) * 0.9;
+                winheight = (unsigned int)(DisplayHeight(disp, scrn) * 0.9);
         }
     }
